category1-assistant-coach/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
gql[httpx]>=3.5.0
anthropic>=0.18.0
python-dotenv>=1.0.0
